+++ lib/services/utils/url/url_utils.dart
@@ -76,6 +76,9 @@ class UrlUtils extends IUrlUtils {
     Uri? nativeUri,
     Uri? universalUri,
   }) async {
+    print(
+      'Navigating deep links. Native: ${nativeUri.toString()}, Universal: ${universalUri.toString()}',
+    );
     LoggerUtil.logger.i(
       'Navigating deep links. Native: ${nativeUri.toString()}, Universal: ${universalUri.toString()}',
     );