+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2.2
+
+- Updated to `walletconnect_flutter_v2` v2.1.2
+- Bug fixes and code changes based on how network errors are handled now
+
 ## 1.2.1
 
 - Updated to `walletconnect_flutter_v2` v2.1.1